requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-bootstrap4==23.1
django-crispy-forms==2.0
django-decouple==2.1
django-multiselectfield==0.1.12
django-rosetta==0.9.9
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
polib==1.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
